move to typescript, and fix initial errors so site at least renders
