+++ package.json
@@ -5,17 +5,20 @@
   "keywords": [],
   "main": "src/index.tsx",
   "dependencies": {
+    "@types/node": "^22.13.1",
     "pako": "2.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-scripts": "^5.0.0"
+    "react-scripts": "^5.0.0",
+    "source-map-loader": "^5.0.0",
+    "ts-loader": "^9.5.2"
   },
   "devDependencies": {
-    "@types/react": "19.0.0",
-    "@types/react-dom": "19.0.0",
+    "@types/react": "^19.0.8",
+    "@types/react-dom": "^19.0.3",
     "loader-utils": "3.2.1",
     "prettier": "3.4.2",
-    "typescript": "5.7.2"
+    "typescript": "^5.7.3"
   },
   "scripts": {
     "start": "react-scripts start",
