+++ src/index.tsx
@@ -1,7 +1,8 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
-import App from "./App";
+import App from "./App.tsx";
+import React from "react";
 
 const rootElement = document.getElementById("root");
 const root = createRoot(rootElement);
